+++ src/client/services/notices.ts
@@ -1,0 +1,7 @@
+import { fetchCall } from './api';
+
+const getNoticeList = () => {};
+
+const getNoticeDetail = () => {};
+
+export { getNoticeList, getNoticeDetail };
