+++ src/shared/components/common/TextEditor/index.tsx
@@ -4,13 +4,19 @@ import { STextEditorWrap, STextArea, STextLength } from './styles';
 interface TextEditorProps {
   text: string;
   maxLength: number;
+  $height?: number;
   onChange: (e: ChangeEvent<HTMLTextAreaElement>) => void;
 }
 
-const TextEditor = ({ text, maxLength, onChange }: TextEditorProps) => {
+const TextEditor = ({
+  text,
+  maxLength,
+  $height = 200,
+  onChange,
+}: TextEditorProps) => {
   return (
     <>
-      <STextEditorWrap>
+      <STextEditorWrap $height={$height}>
         <STextArea
           placeholder="내용을 입력해주세요."
           value={text}
